Reorganize packages, edit BiomeError desc
