+++ src/main/java/org/dimdev/jeid/biome/BiomeError.java
@@ -4,7 +4,7 @@ import net.minecraft.world.biome.BiomeVoid;
 import org.dimdev.jeid.JEID;
 
 public class BiomeError extends BiomeVoid {
-    private static final BiomeError INSTANCE = new BiomeError("error_biome", "A mod doesn't support extended biome IDs -- report to JEID");
+    private static final BiomeError INSTANCE = new BiomeError("error_biome", "A mod doesn't support extended biome IDs -- report to REID");
 
     private BiomeError(String idName, String propName) {
         super(new BiomeProperties(propName));
