+++ src/main/java/org/dimdev/jeid/mixin/core/misc/MixinBlock.java
@@ -1,4 +1,4 @@
-package org.dimdev.jeid.mixin.core;
+package org.dimdev.jeid.mixin.core.misc;
 
 import com.llamalad7.mixinextras.sugar.Local;
 import net.minecraft.block.Block;
