+++ src/main/java/org/dimdev/jeid/mixin/core/misc/client/MixinRenderGlobal.java
@@ -1,4 +1,4 @@
-package org.dimdev.jeid.mixin.core.client;
+package org.dimdev.jeid.mixin.core.misc.client;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.SoundType;
