+++ src/main/java/org/dimdev/jeid/mixin/core/world/MixinBlockStateContainer.java
@@ -1,4 +1,4 @@
-package org.dimdev.jeid.mixin.core;
+package org.dimdev.jeid.mixin.core.world;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.state.IBlockState;
